move vectorEncoding from FieldInfo to FieldEntry
