+++ lucene/codecs/src/java/org/apache/lucene/codecs/simpletext/SimpleTextFieldInfosFormat.java
@@ -29,7 +29,6 @@ import org.apache.lucene.index.IndexFileNames;
 import org.apache.lucene.index.IndexOptions;
 import org.apache.lucene.index.SegmentInfo;
 import org.apache.lucene.index.VectorSimilarityFunction;
-import org.apache.lucene.index.VectorValues.VectorEncoding;
 import org.apache.lucene.store.ChecksumIndexInput;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.store.IOContext;
@@ -69,7 +68,6 @@ public class SimpleTextFieldInfosFormat extends FieldInfosFormat {
   static final BytesRef INDEX_DIM_COUNT = new BytesRef("  index dimensional count ");
   static final BytesRef DIM_NUM_BYTES = new BytesRef("  dimensional num bytes ");
   static final BytesRef VECTOR_NUM_DIMS = new BytesRef("  vector number of dimensions ");
-  static final BytesRef VECTOR_ENCODING = new BytesRef("  vector encoding ");
   static final BytesRef VECTOR_SIMILARITY = new BytesRef("  vector similarity ");
   static final BytesRef SOFT_DELETES = new BytesRef("  soft-deletes ");
 
@@ -158,11 +156,6 @@ public class SimpleTextFieldInfosFormat extends FieldInfosFormat {
         int vectorNumDimensions = Integer.parseInt(readString(VECTOR_NUM_DIMS.length, scratch));
 
         SimpleTextUtil.readLine(input, scratch);
-        assert StringHelper.startsWith(scratch.get(), VECTOR_ENCODING);
-        String encoding = readString(VECTOR_ENCODING.length, scratch);
-        VectorEncoding vectorEncoding = vectorEncoding(encoding);
-
-        SimpleTextUtil.readLine(input, scratch);
         assert StringHelper.startsWith(scratch.get(), VECTOR_SIMILARITY);
         String scoreFunction = readString(VECTOR_SIMILARITY.length, scratch);
         VectorSimilarityFunction vectorDistFunc = distanceFunction(scoreFunction);
@@ -186,7 +179,6 @@ public class SimpleTextFieldInfosFormat extends FieldInfosFormat {
                 indexDimensionalCount,
                 dimensionalNumBytes,
                 vectorNumDimensions,
-                vectorEncoding,
                 vectorDistFunc,
                 isSoftDeletesField);
       }
@@ -207,10 +199,6 @@ public class SimpleTextFieldInfosFormat extends FieldInfosFormat {
 
   public DocValuesType docValuesType(String dvType) {
     return DocValuesType.valueOf(dvType);
-  }
-
-  public VectorEncoding vectorEncoding(String vectorEncoding) {
-    return VectorEncoding.valueOf(vectorEncoding);
   }
 
   public VectorSimilarityFunction distanceFunction(String scoreFunction) {
@@ -307,10 +295,6 @@ public class SimpleTextFieldInfosFormat extends FieldInfosFormat {
 
         SimpleTextUtil.write(out, VECTOR_NUM_DIMS);
         SimpleTextUtil.write(out, Integer.toString(fi.getVectorDimension()), scratch);
-        SimpleTextUtil.writeNewline(out);
-
-        SimpleTextUtil.write(out, VECTOR_ENCODING);
-        SimpleTextUtil.write(out, fi.getVectorEncoding().name(), scratch);
         SimpleTextUtil.writeNewline(out);
 
         SimpleTextUtil.write(out, VECTOR_SIMILARITY);
