+++ lucene/core/src/java/org/apache/lucene/codecs/lucene90/Lucene90FieldInfosFormat.java
@@ -30,7 +30,6 @@ import org.apache.lucene.index.IndexFileNames;
 import org.apache.lucene.index.IndexOptions;
 import org.apache.lucene.index.SegmentInfo;
 import org.apache.lucene.index.VectorSimilarityFunction;
-import org.apache.lucene.index.VectorValues.VectorEncoding;
 import org.apache.lucene.store.ChecksumIndexInput;
 import org.apache.lucene.store.DataOutput;
 import org.apache.lucene.store.Directory;
@@ -192,7 +191,6 @@ public final class Lucene90FieldInfosFormat extends FieldInfosFormat {
                     pointIndexDimensionCount,
                     pointNumBytes,
                     vectorDimension,
-                    VectorEncoding.FLOAT32,
                     vectorDistFunc,
                     isSoftDeletesField);
             infos[i].checkConsistency();
