+++ lucene/core/src/java/org/apache/lucene/index/FieldInfo.java
@@ -18,7 +18,6 @@ package org.apache.lucene.index;
 
 import java.util.Map;
 import java.util.Objects;
-import org.apache.lucene.index.VectorValues.VectorEncoding;
 
 /**
  * Access to the Field Info file that describes document fields and whether or not they are indexed.
@@ -57,7 +56,6 @@ public final class FieldInfo {
 
   // if it is a positive value, it means this field indexes vectors
   private final int vectorDimension;
-  private final VectorEncoding vectorEncoding;
   private final VectorSimilarityFunction vectorSimilarityFunction;
 
   // whether this field is used as the soft-deletes field
@@ -82,7 +80,6 @@ public final class FieldInfo {
       int pointIndexDimensionCount,
       int pointNumBytes,
       int vectorDimension,
-      VectorEncoding vectorEncoding,
       VectorSimilarityFunction vectorSimilarityFunction,
       boolean softDeletesField) {
     this.name = Objects.requireNonNull(name);
@@ -108,7 +105,6 @@ public final class FieldInfo {
     this.pointIndexDimensionCount = pointIndexDimensionCount;
     this.pointNumBytes = pointNumBytes;
     this.vectorDimension = vectorDimension;
-    this.vectorEncoding = vectorEncoding;
     this.vectorSimilarityFunction = vectorSimilarityFunction;
     this.softDeletesField = softDeletesField;
     this.checkConsistency();
@@ -233,10 +229,8 @@ public final class FieldInfo {
     verifySameVectorOptions(
         fieldName,
         this.vectorDimension,
-        this.vectorEncoding,
         this.vectorSimilarityFunction,
         o.vectorDimension,
-        o.vectorEncoding,
         o.vectorSimilarityFunction);
   }
 
@@ -353,25 +347,19 @@ public final class FieldInfo {
   static void verifySameVectorOptions(
       String fieldName,
       int vd1,
-      VectorEncoding ve1,
       VectorSimilarityFunction vsf1,
       int vd2,
-      VectorEncoding ve2,
       VectorSimilarityFunction vsf2) {
-    if (vd1 != vd2 || vsf1 != vsf2 || ve1 != ve2) {
+    if (vd1 != vd2 || vsf1 != vsf2) {
       throw new IllegalArgumentException(
           "cannot change field \""
               + fieldName
               + "\" from vector dimension="
               + vd1
-              + ", vector encoding="
-              + ve1
               + ", vector similarity function="
               + vsf1
               + " to inconsistent vector dimension="
               + vd2
-              + ", vector encoding="
-              + ve2
               + ", vector similarity function="
               + vsf2);
     }
@@ -480,11 +468,6 @@ public final class FieldInfo {
   /** Returns the number of dimensions of the vector value */
   public int getVectorDimension() {
     return vectorDimension;
-  }
-
-  /** Returns the number of dimensions of the vector value */
-  public VectorEncoding getVectorEncoding() {
-    return vectorEncoding;
   }
 
   /** Returns {@link VectorSimilarityFunction} for the field */
