+++ lucene/core/src/java/org/apache/lucene/index/VectorValues.java
@@ -32,23 +32,6 @@ public abstract class VectorValues extends DocIdSetIterator {
   /** The maximum length of a vector */
   public static final int MAX_DIMENSIONS = 1024;
 
-  /** The numeric datatype of the vector values. */
-  public enum VectorEncoding {
-
-    /**
-     * Encodes vector using 8 bits of precision per sample. Use only with DOT_PRODUCT similarity.
-     * NOTE: this can enable significant storage savings and faster searches, at the cost of some
-     * possible loss of precision. In order to use it, all vectors must be of the same norm, as
-     * measured by the sum of the squares of the scalar values, and those values must be in the
-     * range [-128, 127]. This applies to both document and query vectors. Using nonconforming
-     * vectors can result in errors or poor search results.
-     */
-    BYTE,
-
-    /** Encodes vector using 32 bits of precision per sample in IEEE floating point format. */
-    FLOAT32
-  }
-
   /** Sole constructor */
   protected VectorValues() {}
