+++ lucene/core/src/test/org/apache/lucene/index/TestFieldInfos.java
@@ -26,7 +26,6 @@ import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
 import org.apache.lucene.document.FieldType;
 import org.apache.lucene.document.StringField;
-import org.apache.lucene.index.VectorValues.VectorEncoding;
 import org.apache.lucene.store.Directory;
 import org.apache.lucene.tests.analysis.MockAnalyzer;
 import org.apache.lucene.tests.util.LuceneTestCase;
@@ -261,7 +260,6 @@ public class TestFieldInfos extends LuceneTestCase {
               0,
               0,
               0,
-              VectorEncoding.FLOAT32,
               VectorSimilarityFunction.EUCLIDEAN,
               false));
     }
@@ -281,7 +279,6 @@ public class TestFieldInfos extends LuceneTestCase {
                 0,
                 0,
                 0,
-                VectorEncoding.FLOAT32,
                 VectorSimilarityFunction.EUCLIDEAN,
                 false));
     assertEquals("Field numbers 0 through 9 were allocated", 10, idx);
@@ -303,7 +300,6 @@ public class TestFieldInfos extends LuceneTestCase {
                 0,
                 0,
                 0,
-                VectorEncoding.FLOAT32,
                 VectorSimilarityFunction.EUCLIDEAN,
                 false));
     assertEquals("Field numbers should reset after clear()", 0, idx);
