+++ lucene/core/src/test/org/apache/lucene/index/TestPendingSoftDeletes.java
@@ -27,7 +27,6 @@ import org.apache.lucene.document.Document;
 import org.apache.lucene.document.Field;
 import org.apache.lucene.document.NumericDocValuesField;
 import org.apache.lucene.document.StringField;
-import org.apache.lucene.index.VectorValues.VectorEncoding;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.store.ByteBuffersDirectory;
 import org.apache.lucene.store.Directory;
@@ -197,7 +196,6 @@ public class TestPendingSoftDeletes extends TestPendingDeletes {
             0,
             0,
             0,
-            VectorEncoding.FLOAT32,
             VectorSimilarityFunction.EUCLIDEAN,
             true);
     List<Integer> docsDeleted = Arrays.asList(1, 3, 7, 8, DocIdSetIterator.NO_MORE_DOCS);
@@ -235,7 +233,6 @@ public class TestPendingSoftDeletes extends TestPendingDeletes {
             0,
             0,
             0,
-            VectorEncoding.FLOAT32,
             VectorSimilarityFunction.EUCLIDEAN,
             true);
     for (DocValuesFieldUpdates update : updates) {
@@ -298,7 +295,6 @@ public class TestPendingSoftDeletes extends TestPendingDeletes {
             0,
             0,
             0,
-            VectorEncoding.FLOAT32,
             VectorSimilarityFunction.EUCLIDEAN,
             true);
     List<Integer> docsDeleted = Arrays.asList(1, DocIdSetIterator.NO_MORE_DOCS);
@@ -366,7 +362,6 @@ public class TestPendingSoftDeletes extends TestPendingDeletes {
             0,
             0,
             0,
-            VectorEncoding.FLOAT32,
             VectorSimilarityFunction.EUCLIDEAN,
             true);
     List<DocValuesFieldUpdates> updates =
@@ -403,7 +398,6 @@ public class TestPendingSoftDeletes extends TestPendingDeletes {
             0,
             0,
             0,
-            VectorEncoding.FLOAT32,
             VectorSimilarityFunction.EUCLIDEAN,
             true);
     updates = Arrays.asList(singleUpdate(Arrays.asList(1, DocIdSetIterator.NO_MORE_DOCS), 3, true));
