+++ lucene/test-framework/src/java/org/apache/lucene/tests/index/BaseIndexFileFormatTestCase.java
@@ -360,7 +360,6 @@ abstract class BaseIndexFileFormatTestCase extends LuceneTestCase {
             proto.getPointIndexDimensionCount(),
             proto.getPointNumBytes(),
             proto.getVectorDimension(),
-            proto.getVectorEncoding(),
             proto.getVectorSimilarityFunction(),
             proto.isSoftDeletesField());
