+++ lucene/test-framework/src/java/org/apache/lucene/tests/index/RandomPostingsTester.java
@@ -63,7 +63,6 @@ import org.apache.lucene.index.TermState;
 import org.apache.lucene.index.Terms;
 import org.apache.lucene.index.TermsEnum;
 import org.apache.lucene.index.VectorSimilarityFunction;
-import org.apache.lucene.index.VectorValues.VectorEncoding;
 import org.apache.lucene.internal.tests.IndexPackageAccess;
 import org.apache.lucene.internal.tests.TestSecrets;
 import org.apache.lucene.search.DocIdSetIterator;
@@ -164,7 +163,6 @@ public class RandomPostingsTester {
               0,
               0,
               0,
-              VectorEncoding.FLOAT32,
               VectorSimilarityFunction.EUCLIDEAN,
               false);
       fieldUpto++;
@@ -736,7 +734,6 @@ public class RandomPostingsTester {
               0,
               0,
               0,
-              VectorEncoding.FLOAT32,
               VectorSimilarityFunction.EUCLIDEAN,
               false);
     }
